Add a (fairly simplistic) child model system to NotesModel
This should make it easier to use NotesModels for more funky purposes,
and just generally allow us to not have to pull the objects out by
force, and instead be able to just use a model for any columny purposes
as well. Hopefully help make our playgrids' code look more pleasant and
simple.