Sequences should be advanced by the current cumulative beat +1
Otherwise we'll end up scheduling things for the current beat, which
will not work, as that will be immediately ignored by moving ahead. So,
schedule things for the /next/ beat, and suddenly happiness.

Also a few bits of cosmetic and commentary fixup (and a little
belts-and-braces to try and make sure we're not dereffing a potentially
delayed syncTimer variable... might still happen, but really don't want
to do this in the functions that need that variable)