Switch to the positionBuffer logic (may cause issues - needs testing!)
The issues this may cause are due to it not currently using the delay
adjustment stuff, which we very possibly still need. So... Needs testing
and probably adjustment, but it's a step in the right direction, as it
gives us much more expressive power (including delay start and explicit
duration values for notes, and hooks for the ability to do other stuff
like setting looping per-note, note-repeat). Still need to work out how
to make arbitrary commands fit into this, but might have an idea.