PatternModel: Add thumbnailUrl, trackIndex, partIndex, partName properties
thumbnailUrl is a convenience generator for the current bank's thumbnail
(you can still manually build them without that, but this makes a lot of
general-purpose operations much simpler).

trackIndex, partIndex, and partName are used by the part system in
zynthbox to work out what patterns belong to which tracks, and which
part they represent inside that track.

Also make use of the long-operation thing to make sure we're not
updating stuff unless we need to.