SequenceModel: Pull more logic from javascript to c++, fix up some load/save issues
The loading issues are a bit mixed, and include things like being more
strict about handling parenting of what we're doing and the like. No
particular reason not to pass the sequence itself in when requesting the
pattern models, and then handling insertion ourselves. That sort of
stuff. Also make sure to warn things are going to take a while (which
paradoxically helps make things not take so long).

Also only save if we're dirty. And don't just give ourselves a name if
we're not associated with a song yet. This might cause issues somewhere,
but it makes some things much faster as well.

We're now loading an empty song in about seven seconds. Can still be
faster, but is pretty good from about a minute a pop, and doing it
twice. So, good progress there, i feel. Hopefully this didn't blow
anything else up, so next up: work out if we broke something.