SequenceModel: Reset patterns, don't just clear their notes
When loading a pattern and filling the empty spaces, we don't want any
of the data left, not just the notes. We do want to retain the internal
bits like name and connections to tracks and the like, though, so the
resetPattern function only resets those things that actually make up a
pattern's internal musical logic setup, not our housekeeping bits.