PlayGridManager: Make it possible to fetch an un-loaded Sequence
This isn't really a general-use thing, but it helps us not unnecessarily
perform a bunch of pattern operations when we're just going to change
that immediately afterwards anyway for the scene sequences by setting
the song on them. Shaves a touch under two seconds off of the loading
time, which is always nice.