PatternModel: Consume midi events for samplersynth destination
This should help us bring the sampler more in sync with the other
synths. There is still going to be some small delay, due to how the midi
notes get passed in, but that is something we'll want to fix anyway, but
elsewhere, and this will then simply end up better as that happens. (in
short, need to move away from rtmidi)