PatternModel: Honour chainedSounds, pass to ZLRouter (requires new libzl)
This is the final bit of glue that should once and for all make the
chained sounds stuff in zynthian work with our router, in a way that
doesn't constantly break when anything other than the default base stuff
is picked. In short: ZLRouter will now actually route Zynthian notes to
the channel that Zynthian expects to receive them on.