SequenceModel: Experimentally trust synctimer to schedule ahead
Now that SyncTimer handles running ahead of the jack playhead, we should
be able to actually do some more clever stuff with this. First step
here, but we should, with this, be able to switch the current part on a
step and have it work as expected. Conceptually, at least, but this
should basically work. Still need to prefill, to be sure, but...