PatternModel: Implement lookahead, so negative delays can be applied
Also adds a function for invalidating positions, as previous positions
need invalidating now as well. Seems to work decently, maybe we want to
do some analysis on how heavy this actually is. If it's heavy, we could
match up invalidation and position generating, and generate on-load and
on-set instead. Or maybe it's not a problem and this is fine. Needs
perf/hotspot-ing, just in general.