PlayGridManager: Simplify the most-recent-note handler
Previously it would try and filter out repeat notes, but since we're now
controlling what comes into the function more tightly, we actually
/want/ duplicates if they arrive, so let's do this.

Also, because apparently we end up counting incorrectly sometimes, and
it annoys me greatly that i don't know why, just set off if we receive
an off. It's incorrect, but removes hanging notes. Not cool, this needs
deeper investigation, but it seems also to be entirely visual (as sounds
go away just fine), so... yeah, not sure what's going on there.