NotesModel: Add persisted keyed metadata to positions
This will allow us to add metadata using a key/value system to each
position. This is the ground work needed to allow for the introduction
of position-located non-note commands and the like for our PatternModel,
for example allowing us to send arbitrary midi events, pitch changes,
all that sort of fun. Just the initial stuff, but needed for that to be
possible.