SegmentHandler: Explicitly progress from PGM, and increase playhead first
This is a reversal of the playhead progress logic from an earlier
commit, but with reworks in other places, this is what we actually want.
With this change. This, combined with the SegmentHandler-first
progression logic means that we now in fact have the first note of a
pattern included, which we do kind of want.