PatternModel: Add (untested) live-recording logic
This is intended to allow us to record notes during playback. That is,
start playback, turn on live recording, and any notes recognised as
being for the pattern will be added to the current step (when the
matching note-off message arrives, so we also capture the duration).

Untested at this point, and will likely need adjustment, so next step is
to add a way to turn it off in drumatique.