PatternModel: Live midi recording seems to be working now
Also adjusted the logic so it should be very light on resources. There's
still a 'new' in there that i'd quite like to get rid of, and also the
other bits in the handleMidiMessage thing that would be good to get
sorted some, so it's not creating a new couple of juce containers each
time, which... seems really, really pricy for what that function is
doing. But, live recording is now creating new subnotes for anything
recorded, and setting the values if they're different, and not if
they're not, so at this point seems pretty performant. Needs ui support
to actually work, initial implementation incoming for that shortly.