PatternModel: Disable live recording if user switches to another pattern
The logic here being that if the user switches to a different pattern,
we want to try to avoid doing changes they can't directly observe. There
are use cases where you might want to do otherwise, but we also want to
be friendly to newbies and avoid surprises, so trying to achieve a
"anything not visible is immutable" type situation seems not the worst
idea.