PatternModel: Add zl.dashboard property, so we can handle current track
This fixes the initial-state fault where we would end up just picking
the last track as the current one, resulting in us being out of sync
with the state of zl. This now removes the internal state, and we will
always use the zl.dashboard current track to set the current pattern.
This furthers the whole "don't keep state in the components" concept
that we've been moving towards, and we're getting very close to having
that entirely sorted. Famous last words, there, but there's less and
less of this sort of thing happening.