MidiListener: Port from RtMidi to jack, and actually gain timestamps
This removes the old RtMidi based listening in favour of something based
on jack. The immediate result is that we are able to add timestamps (we
could make them more precise, but initially we get that), and further no
longer depend on RtMidi, which is great for casual use, but not great
when precision is required.